pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "avatar-engine"
version = "0.1.0"
description = "AI assistant integration library for Claude Code and Gemini CLI"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "Antonin Stefanutti" }
]
keywords = ["ai", "claude", "gemini", "assistant", "avatar", "llm", "agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "pyyaml>=6.0",
    "acp>=0.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "mypy>=1.0",
    "ruff>=0.1",
]
cli = [
    "click>=8.0",
    "rich>=13.0",
]

[project.scripts]
avatar = "avatar_engine.cli:cli"

[project.urls]
Homepage = "https://github.com/raven2cz/avatar-engine"
Repository = "https://github.com/raven2cz/avatar-engine"
Issues = "https://github.com/raven2cz/avatar-engine/issues"

[tool.hatch.build.targets.wheel]
packages = ["avatar_engine"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
# Exclude integration tests from default runs
addopts = "--ignore=tests/integration"
markers = [
    "integration: marks tests as integration tests (require real CLI tools)",
    "gemini: marks tests as requiring Gemini CLI",
    "claude: marks tests as requiring Claude CLI",
    "slow: marks tests as slow (API calls)",
]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
